requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
seaborn>=0.12.0
umap-learn>=0.5.3
Pillow>=9.4.0
ipykernel>=6.0.0
jupyter>=1.0.0
